sample-chatbot-history/requirements.txt
langchain
python-dotenv
langchain_community
langchain-text-splitters
langchain-openai
langchain-chatgroq
langchain-groq
